--- cypress/support/e2e.js
@@ -1,2 +1,0 @@
-import 'cypress-hmr-restarter'
-import '@testing-library/cypress/add-commands'
