+++ src/components/list-item-list.js
@@ -1,41 +1,6 @@
-/** @jsx jsx */
-import {jsx} from '@emotion/core'
+// no final
 
-import {useListItems} from 'utils/list-items'
-import {BookListUL} from './lib'
-import {BookRow} from './book-row'
-import {Profiler} from './profiler'
+export * from './list-item-list.exercise'
 
-function ListItemList({filterListItems, noListItems, noFilteredListItems}) {
-  const listItems = useListItems()
-
-  const filteredListItems = listItems.filter(filterListItems)
-
-  if (!listItems.length) {
-    return <div css={{marginTop: '1em', fontSize: '1.2em'}}>{noListItems}</div>
-  }
-  if (!filteredListItems.length) {
-    return (
-      <div css={{marginTop: '1em', fontSize: '1.2em'}}>
-        {noFilteredListItems}
-      </div>
-    )
-  }
-
-  return (
-    <Profiler
-      id="List Item List"
-      metadata={{listItemCount: filteredListItems.length}}
-    >
-      <BookListUL>
-        {filteredListItems.map(listItem => (
-          <li key={listItem.id} aria-label={listItem.book.title}>
-            <BookRow book={listItem.book} />
-          </li>
-        ))}
-      </BookListUL>
-    </Profiler>
-  )
-}
-
-export {ListItemList}
+// 💯 Production Monitoring
+// export * from './list-item-list.extra-3'
