+++ src/index.exercise.js
@@ -1,0 +1,19 @@
+import {loadDevTools} from './dev-tools/load'
+import './bootstrap'
+import * as React from 'react'
+import {createRoot} from 'react-dom/client'
+import {App} from './app'
+import {AppProviders} from './context'
+
+// ignore the rootRef in this file. I'm just doing it here to make
+// the tests I write to check your work easier.
+export const rootRef = {}
+loadDevTools(() => {
+  const root = createRoot(document.getElementById('root'))
+  root.render(
+    <AppProviders>
+      <App />
+    </AppProviders>,
+  )
+  rootRef.current = root
+})
