+++ src/index.js
@@ -1,17 +1,6 @@
-import {loadDevTools} from './dev-tools/load'
-import './bootstrap'
-import * as React from 'react'
-import {createRoot} from 'react-dom/client'
-import {Profiler} from 'components/profiler'
-import {App} from './app'
-import {AppProviders} from './context'
+// no final
 
-loadDevTools(() => {
-  createRoot(document.getElementById('root')).render(
-    <Profiler id="App Root" phases={['mount']}>
-      <AppProviders>
-        <App />
-      </AppProviders>
-    </Profiler>,
-  )
-})
+export * from './index.exercise'
+
+// 💯 Production Monitoring
+// export * from './index.extra-3'
