+++ src/utils/hooks.js
@@ -58,7 +58,7 @@ function useAsync(initialState) {
         },
         error => {
           setError(error)
-          return Promise.reject(error)
+          return error
         },
       )
     },
